requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.5
pydantic==2.9.0
python-multipart==0.0.9
python-docx==1.1.2
pdfplumber==0.11.4
httpx==0.27.2
openai==1.52.0
pytest==8.3.3
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1
beautifulsoup4==4.12.3
lxml==5.3.0
reportlab==4.4.5
slowapi==0.1.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
faiss-cpu>=1.9.0  # For RAG vector database (optional)


